Split Device Context and ContextProvider into 2 files for react fast refresh
