+++ src/App.tsx
@@ -1,7 +1,7 @@
 import Header from "./components/Header";
 import Content from "./components/Content";
 import Footer from "./components/Footer";
-import { ProvideDevice } from "./contexts/Device";
+import { ProvideDevice } from "./contexts/DeviceProvider";
 import "./styles/App.scss";
 
 export default function App() {
